app/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "job-tracker"
version = "0.1.0"
description = "A CLI tool to track job applications."
authors = [{name = "Michael Liz", email = "Lmic1219@gmail.com"}]
dependencies = [
    "click",
    "psycopg2-binary",
    "python-dotenv"
]
requires-python = ">=3.8"

[project.scripts]
jt = "cli:cli"
